+++ libs/tools.py
@@ -18,6 +18,7 @@ import re
 import ruamel.yaml
 from ruamel.yaml.error import YAMLError
 
+# Debug
 # from pdb import set_trace as st
 
 LOGGER = logging.getLogger('aws-tower')
@@ -380,6 +381,7 @@ def search_filter_in(asset, filter_str):
     - port:xxx -> asset.security_groups (ELB, EC2)
     - engine:xxx -> asset.engine (RDS)
     - version:xxx -> asset.version (EKS, RDS)
+    - os:xxx -> asset.version (EC2)
     """
     filter_str = filter_str.lower()
     if asset is None:
@@ -390,20 +392,23 @@ def search_filter_in(asset, filter_str):
         for security_group in asset.security_groups:
             is_found |= port in asset.security_groups[security_group].keys()
     elif filter_str.startswith('engine:') and hasattr(asset, 'engine'):
-        is_found = asset.engine.startswith(filter_str.split(':')[1])
+        is_found = asset.engine.lower().startswith(filter_str.split(':')[1])
     elif filter_str.startswith('version:'):
         version = filter_str.split(':')[1]
         if asset.get_type() == 'EKS':
             is_found = asset.version.startswith(version)
         if asset.get_type() == 'RDS' and '==' in asset.engine:
             is_found = asset.engine.split('==')[1].startswith(version)
+    elif filter_str.startswith('os:') and hasattr(asset, 'operating_system'):
+        is_found = asset.operating_system.lower().startswith(filter_str.split(':')[1])
     else:
         if filter_str in asset.name.lower():
             return True
         for attribute in [
             'aliases', 'api_endpoint', 'arn',
             'dns_record', 'dst_account_id', 'endpoint',
-            'engine', 'private_ip', 'public_ip', 'src_account_id', 'url']:
+            'engine', 'private_ip', 'public_ip', 'src_account_id', 'url',
+            'role_poweruser', 'role_admin']:
             is_found |= hasattr(asset, attribute) and \
                 isinstance(getattr(asset, attribute), str) and \
                 filter_str in getattr(asset, attribute).lower()