+++ CHANGELOG.md
@@ -4,6 +4,19 @@ CHANGELOG
 AWS-TOWER
 -----
 
+4.4.2
+-----
+
+2023/01/26
+
+### New feature
+  - Add 'os' filter
+  - Add 'MQ' asset type
+
+### Changes
+  - Display InstanceID for EC2
+  - Filter IAM roles to display only Instance Profiles
+
 4.4.1
 -----